refactor(package.json): :art: 格式化代码
