+++ packages/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from "./useRefTemplate"
+export * from "./src/useRefTemplate"
