+++ packages/hooks/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@squid-element/hooks",
   "version": "0.0.0",
+  "main": "index.ts",
   "license": "MIT",
   "peerDependencies": {
     "vue": "^3.0.3"
