+++ packages/icon/src/index.vue
@@ -13,6 +13,7 @@
 import { computed, CSSProperties, defineComponent, toRefs } from "vue"
 import { props as componentProps } from "./props"
 import { createFromIconfontCN } from "@ant-design/icons-vue"
+
 const IconFont: any = createFromIconfontCN({
   // 在 iconfont.cn 上生成
   // TODO: cdn后续在取得项目配置中获取
